+++ test/test_helper.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "i18n"
+
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
 ENV["RAILS_ENV"] = "test"